refactor: 배팅 인원 제한 기능 이동
